+++ Book Master Backend/src/main/java/com/csc340/demo/Review/ReviewRepository.java
@@ -3,6 +3,7 @@
 
 package com.csc340.demo.Review;
 
+import com.csc340.demo.Book.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
@@ -13,7 +14,7 @@ import java.util.List;
 public interface ReviewRepository extends JpaRepository<Review, Long> {
 
     @Query(value = "select * from reviews r where r.book_id = ?1", nativeQuery = true)
-    List<Review> findByBookId(Long bookId);
+    List<Review> findByBookId(int bookId);
 
     @Query(value = "SELECT COUNT(*) FROM reviews WHERE book_id = :bookId", nativeQuery = true)
     long countReviews(@Param("bookId") int bookId);
